+++ lovelyshades/src/main/java/com/lovelyshades/repository/DetalleFacturaRepository.java
@@ -3,5 +3,5 @@ package com.lovelyshades.repository;
 import com.lovelyshades.entity.DetalleFactura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalleFacturaRepository extends JpaRepository<DetalleFactura, Long> {
+public interface DetalleFacturaRepository extends JpaRepository<DetalleFactura, Integer> {
 }
