+++ lovelyshades/src/main/java/com/lovelyshades/repository/FacturaRepository.java
@@ -3,5 +3,5 @@ package com.lovelyshades.repository;
 import com.lovelyshades.entity.Factura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaRepository extends JpaRepository<Factura, Long> {
+public interface FacturaRepository extends JpaRepository<Factura, Integer> {
 }
