+++ lovelyshades/src/main/java/com/lovelyshades/repository/InventarioRepository.java
@@ -3,5 +3,5 @@ package com.lovelyshades.repository;
 import com.lovelyshades.entity.Inventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventarioRepository extends JpaRepository<Inventario, Long> {
+public interface InventarioRepository extends JpaRepository<Inventario, Integer> {
 }
