+++ lovelyshades/src/main/java/com/lovelyshades/repository/ProductoRepository.java
@@ -3,5 +3,5 @@ package com.lovelyshades.repository;
 import com.lovelyshades.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Long> {
+public interface ProductoRepository extends JpaRepository<Producto, Integer> {
 }
