+++ lovelyshades/src/main/java/com/lovelyshades/service/CrudService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface CrudService<T> {
     List<T> listar();
-    T obtenerPorId(Long id);
+    T obtenerPorId(Integer id);
     T crear(T entity);
-    T actualizar(Long id, T entity);
-    void eliminar(Long id);
+    T actualizar(Integer id, T entity);
+    void eliminar(Integer id);
 }
